# core/vm: use fixed uint256 library instead of big