--- a/core/vm/eips.go
+++ b/core/vm/eips.go
@@ -61,7 +61,8 @@ func enable1884(jt *JumpTable) {
 }
 
 func opSelfBalance(pc *uint64, interpreter *EVMInterpreter, contract *Contract, memory *Memory, stack *Stack) ([]byte, error) {
-	balance := interpreter.intPool.get().Set(interpreter.evm.StateDB.GetBalance(contract.Address()))
+	balance := interpreter.intPool.get()
+	balance.SetFromBig(interpreter.evm.StateDB.GetBalance(contract.Address()))
 	stack.push(balance)
 	return nil, nil
 }
@@ -81,7 +82,8 @@ func enable1344(jt *JumpTable) {
 
 // opChainID implements CHAINID opcode
 func opChainID(pc *uint64, interpreter *EVMInterpreter, contract *Contract, memory *Memory, stack *Stack) ([]byte, error) {
-	chainId := interpreter.intPool.get().Set(interpreter.evm.chainConfig.ChainID)
+	chainId := interpreter.intPool.get()
+	chainId.SetFromBig(interpreter.evm.chainConfig.ChainID)
 	stack.push(chainId)
 	return nil, nil
 }