--- a/core/vm/intpool.go
+++ b/core/vm/intpool.go
@@ -19,6 +19,8 @@ package vm
 import (
 	"math/big"
 	"sync"
+
+	"github.com/holiman/uint256"
 )
 
 var checkVal = big.NewInt(-42)
@@ -26,7 +28,7 @@ var checkVal = big.NewInt(-42)
 const poolLimit = 256
 
 // intPool is a pool of big integers that
-// can be reused for all big.Int operations.
+// can be reused for all uint256.Int operations.
 type intPool struct {
 	pool *Stack
 }
@@ -37,25 +39,25 @@ func newIntPool() *intPool {
 
 // get retrieves a big int from the pool, allocating one if the pool is empty.
 // Note, the returned int's value is arbitrary and will not be zeroed!
-func (p *intPool) get() *big.Int {
+func (p *intPool) get() *uint256.Int {
 	if p.pool.len() > 0 {
 		return p.pool.pop()
 	}
-	return new(big.Int)
+	return new(uint256.Int)
 }
 
 // getZero retrieves a big int from the pool, setting it to zero or allocating
 // a new one if the pool is empty.
-func (p *intPool) getZero() *big.Int {
+func (p *intPool) getZero() *uint256.Int {
 	if p.pool.len() > 0 {
-		return p.pool.pop().SetUint64(0)
+		return p.pool.pop().Clear()
 	}
-	return new(big.Int)
+	return new(uint256.Int)
 }
 
 // put returns an allocated big int to the pool to be later reused by get calls.
 // Note, the values as saved as is; neither put nor get zeroes the ints out!
-func (p *intPool) put(is ...*big.Int) {
+func (p *intPool) put(is ...*uint256.Int) {
 	if len(p.pool.data) > poolLimit {
 		return
 	}
@@ -63,7 +65,7 @@ func (p *intPool) put(is ...*big.Int) {
 		// verifyPool is a build flag. Pool verification makes sure the integrity
 		// of the integer pool by comparing values to a default value.
 		if verifyPool {
-			i.Set(checkVal)
+			i.SetFromBig(checkVal)
 		}
 		p.pool.push(i)
 	}