… core/vm/instructions.go
… go.sum