@chfast 2020-01-27
- Eliminating so many memory allocations is big win nevertheless.
- Does uint256 require any memory allocations?
- Are you able to measure total memory footprint of the node?
- In the final version the `intPool` should be eliminated, is that correct?