@holiman 2020-01-27
>    Eliminating so many memory allocations is big win nevertheless.

Yep, I agree

>    Does uint256 require any memory allocations?

Well, the struct consists of `4 x uint64` internally, but for normal operations it does not allocate.  There are exceptions, see https://github.com/holiman/uint256/#benchmarks , some examples
```
Benchmark_Exp/large/uint256-6            	  100000	     17744 ns/op	      32 B/op	       1 allocs/op
Benchmark_Div/large/uint256-6            	 5000000	       241 ns/op	     128 B/op	       3 allocs/op
Benchmark_MulMod/large/uint256-6         	 1000000	      1245 ns/op	     608 B/op	      11 allocs/op
Benchmark_SDiv/large/uint256-6           	 5000000	       255 ns/op	     128 B/op	       3 allocs/op
```
(`MulMod` is the worst, it uses `big.Int` behind the scenes). 

>    Are you able to measure total memory footprint of the node?

Yep, we'll see that in the charts, and also the number of allocs and the gc workload. 

>    In the final version the intPool should be eliminated, is that correct?

Yes, I would assume so, but we'll have to benchmark that change. So first `big.Int` vs `uint256`, then (possibly) `uint256` versus `uint256-unpooled`.