@holiman 2020-01-28
I swapped this in during the execution of an earlier full-sync - the down-spike is when the I swapped the code. No huge gains
![Screenshot_2020-01-28 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/73290235-5f06c800-41fe-11ea-9e3d-94f6fef816c1.png)