@holiman 2020-03-19
Updated with the latest improvements from @chfast 
```
[user@work vm]$ benchstat before_pawel.txt after_pawel.txt 
name            old time/op  new time/op   delta
OpAdd64-6       96.0ns ± 3%  100.1ns ±22%     ~     (p=0.841 n=5+5)
OpAdd128-6       117ns ± 3%    113ns ± 4%     ~     (p=0.087 n=5+5)
OpAdd256-6       193ns ±19%    178ns ± 2%   -8.09%  (p=0.000 n=5+4)
OpSub64-6        113ns ±15%    102ns ± 1%   -9.57%  (p=0.016 n=5+4)
OpSub128-6       131ns ± 3%    129ns ± 3%     ~     (p=0.127 n=5+5)
OpSub256-6       183ns ± 2%    184ns ± 3%     ~     (p=0.794 n=4+5)
OpMul-6          251ns ±19%    205ns ±16%     ~     (p=0.056 n=5+5)
OpDiv256-6       427ns ± 5%    289ns ± 2%  -32.41%  (p=0.008 n=5+5)
OpDiv128-6       317ns ± 2%    175ns ± 1%  -44.65%  (p=0.016 n=5+4)
OpDiv64-6        119ns ±22%    118ns ±25%     ~     (p=0.460 n=5+5)
OpSdiv-6         444ns ± 1%    306ns ± 2%  -31.06%  (p=0.029 n=4+4)
OpMod-6          211ns ±23%    189ns ± 2%  -10.16%  (p=0.024 n=5+5)
OpSmod-6         217ns ± 3%    212ns ± 1%   -2.24%  (p=0.016 n=5+4)
OpExp-6         19.6µs ± 1%    9.4µs ± 1%  -52.22%  (p=0.016 n=4+5)
OpSignExtend-6   196ns ±20%    187ns ±15%     ~     (p=0.198 n=5+5)
OpLt-6           184ns ± 2%    192ns ±15%     ~     (p=0.738 n=5+5)
OpGt-6           197ns ±21%    191ns ±13%     ~     (p=0.690 n=5+5)
OpSlt-6          196ns ± 3%    199ns ± 3%     ~     (p=0.286 n=4+5)
OpSgt-6          198ns ± 3%    197ns ± 0%     ~     (p=0.635 n=5+4)
OpEq-6           186ns ± 4%    189ns ± 9%     ~     (p=1.000 n=5+5)
OpEq2-6          193ns ±19%    186ns ± 9%     ~     (p=0.389 n=5+5)
OpAnd-6          187ns ± 2%    184ns ± 1%     ~     (p=0.079 n=5+4)
OpOr-6           182ns ± 1%    181ns ± 1%     ~     (p=0.476 n=5+4)
OpXor-6          183ns ± 1%    188ns ± 5%     ~     (p=0.603 n=4+5)
OpByte-6         197ns ±23%    196ns ±22%     ~     (p=0.730 n=5+5)
OpAddmod-6       651ns ±29%    421ns ±15%  -35.23%  (p=0.008 n=5+5)
OpMulmod-6      1.32µs ± 4%   0.54µs ± 1%  -58.89%  (p=0.016 n=5+4)
OpSHL-6          135ns ± 1%    145ns ± 8%   +7.21%  (p=0.016 n=4+5)
OpSHR-6          136ns ± 0%    144ns ± 7%   +5.98%  (p=0.016 n=4+5)
OpSAR-6          143ns ± 7%    149ns ±26%     ~     (p=1.000 n=5+5)
OpIsZero-6      93.4ns ± 2%   93.6ns ± 5%     ~     (p=0.841 n=4+5)
OpMstore-6      95.5ns ± 7%   92.8ns ± 6%     ~     (p=0.222 n=5+5)
OpSHA3-6         798ns ± 2%    784ns ± 1%     ~     (p=0.127 n=5+4)
```